+++ anyblok/tests/test_migration.py
@@ -283,7 +283,7 @@ class TestMigration(TestCase):
         Test = self.registry.Test
         t.check('test').add(Test.other != 'test')
         # particuliar case of check constraint
-        t.check('ck_test__test').drop()
+        t.check('anyblok_ck_test__test').drop()
 
     def test_detect_under_noautocommit_flag(self):
         with self.cnx() as conn:
@@ -481,14 +481,14 @@ class TestMigration(TestCase):
     def test_detect_drop_anyblok_index(self):
         with self.cnx() as conn:
             conn.execute(
-                """CREATE INDEX ix_test__other ON test (other);""")
+                """CREATE INDEX anyblok_ix_test__other ON test (other);""")
         report = self.registry.migration.detect_changed()
         self.assertTrue(
-            report.log_has("Drop index ix_test__other on test"))
+            report.log_has("Drop index anyblok_ix_test__other on test"))
         report.apply_change()
         report = self.registry.migration.detect_changed()
         self.assertFalse(
-            report.log_has("Drop index ix_test__other on test"))
+            report.log_has("Drop index anyblok_ix_test__other on test"))
 
     def test_detect_drop_index_with_reinit_indexes(self):
         with self.cnx() as conn:
@@ -580,7 +580,7 @@ class TestMigration(TestCase):
                 """CREATE TABLE test(
                     integer INT PRIMARY KEY NOT NULL,
                     other CHAR(64)
-                        CONSTRAINT fk_test__other_on_system_blok__name
+                        CONSTRAINT anyblok_fk_test__other_on_system_blok__name
                         references system_blok(name)
                 );""")
         report = self.registry.migration.detect_changed()
@@ -680,15 +680,15 @@ class TestMigration(TestCase):
             conn.execute(
                 """CREATE TABLE test(
                     integer INT PRIMARY KEY NOT NULL,
-                    other CHAR(64) CONSTRAINT uq_test__other UNIQUE
+                    other CHAR(64) CONSTRAINT anyblok_uq_test__other UNIQUE
                 );""")
         report = self.registry.migration.detect_changed()
         self.assertTrue(
-            report.log_has("Drop constraint uq_test__other on test"))
+            report.log_has("Drop constraint anyblok_uq_test__other on test"))
         report.apply_change()
         report = self.registry.migration.detect_changed()
         self.assertFalse(
-            report.log_has("Drop constraint uq_test__other on test"))
+            report.log_has("Drop constraint anyblok_uq_test__other on test"))
 
     def test_detect_drop_unique_constraint_with_reinit_constraints(self):
         with self.cnx() as conn:
@@ -733,11 +733,11 @@ class TestMigration(TestCase):
                 );""")
         report = self.registry.migration.detect_changed()
         self.assertTrue(report.log_has(
-            "Add check constraint ck_testcheck__test on testcheck"))
+            "Add check constraint anyblok_ck_testcheck__test on testcheck"))
         report.apply_change()
         report = self.registry.migration.detect_changed()
         self.assertFalse(report.log_has(
-            "Add check constraint ck_testcheck__test on testcheck"))
+            "Add check constraint anyblok_ck_testcheck__test on testcheck"))
 
     def test_detect_drop_check_constraint(self):
         with self.cnx() as conn:
@@ -761,16 +761,16 @@ class TestMigration(TestCase):
             conn.execute(
                 """CREATE TABLE test(
                     integer INT PRIMARY KEY NOT NULL,
-                    other CHAR(64) CONSTRAINT ck_test__check
+                    other CHAR(64) CONSTRAINT anyblok_ck__test__check
                         CHECK (other != 'test')
                 );""")
         report = self.registry.migration.detect_changed()
         self.assertTrue(report.log_has(
-            "Drop check constraint ck_test__check on test"))
+            "Drop check constraint anyblok_ck__test__check on test"))
         report.apply_change()
         report = self.registry.migration.detect_changed()
         self.assertFalse(report.log_has(
-            "Drop check constraint ck_test__check on test"))
+            "Drop check constraint anyblok_ck__test__check on test"))
 
     def test_detect_drop_check_constraint_with_reinit_constraint(self):
         with self.cnx() as conn: